+++ pkg/cloudbroker/compute/models.go
@@ -435,7 +435,7 @@ type ItemDisk struct {
 	SizeMax uint64 `json:"sizeMax"`
 
 	// Size used
-	SizeUsed uint64 `json:"sizeUsed"`
+	SizeUsed float64 `json:"sizeUsed"`
 
 	// List detailed snapshots
 	Snapshots ListDetailedSnapshots `json:"snapshots"`
